Revert some temporary changes

+++ core/src/types/map.rs
@@ -43,8 +43,8 @@ impl Map {
     }
 
     pub fn add_light(&mut self, amount: i32) {
-        let new_light = self.light + amount as i16;
-        self.light = new_light.clamp(0, i16::MAX); // TODO: What is the actual max light value?
+        let new_light = self.light as i32 + amount;
+        self.light = new_light.clamp(0, i16::MAX as i32) as i16;
     }
 
     pub fn is_sane_coordinates(x: usize, y: usize) -> bool {
